# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Food Delivery API
  version: 1.0.0
  description: |
    REST API for Food Delivery Platform

    This API provides endpoints for managing orders, products, restaurants, and users.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns detailed health status of the service
      tags:
        - Health
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  service:
                    type: string
                    example: "foodie-backend"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2024-01-01T00:00:00Z"

  /ping:
    get:
      summary: Ping endpoint
      description: Simple ping/pong endpoint for quick health checks
      tags:
        - Health
      responses:
        "200":
          description: Server is responding
          content:
            application/json:
              schema:
                type: object
                properties:
                  pong:
                    type: string
                    format: date-time
                    example: "2024-01-01T00:00:00Z"

  /orders:
    get:
      summary: List orders
      tags:
        - Orders
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum:
              [
                pending,
                confirmed,
                preparing,
                out_for_delivery,
                delivered,
                cancelled,
              ]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      summary: Create a new order
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request
        "500":
          description: Internal server error

  /orders/{id}:
    get:
      summary: Get order by ID
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found

  /products:
    get:
      summary: List products
      tags:
        - Products
      parameters:
        - name: restaurant_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
          example: "order-123"
        user_id:
          type: string
          example: "user-456"
        status:
          type: string
          enum:
            [
              pending,
              confirmed,
              preparing,
              out_for_delivery,
              delivered,
              cancelled,
            ]
          example: "pending"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - status
        - created_at

    CreateOrderRequest:
      type: object
      properties:
        user_id:
          type: string
        restaurant_id:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: string
              quantity:
                type: integer
      required:
        - user_id
        - restaurant_id
        - items

    Product:
      type: object
      properties:
        id:
          type: string
        restaurant_id:
          type: string
        name:
          type: string
        price:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
      required:
        - id
        - restaurant_id
        - name
        - price